Move test_all_preview_examples to helper
